+++ src/type/types.ts
@@ -1,0 +1,4 @@
+export type ThemeToggleProps = {
+  theme: "dark" | "light";
+  onToggleTheme: () => void;
+};
